requirements.txt
altair==5.5.0
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.7.0
appnope==0.1.4
asttokens==3.0.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
executing==2.1.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain-anthropic==0.3.0
langchain-core==0.3.27
langgraph==0.2.60
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.48
langsmith==0.2.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.1.0
narwhals==1.18.4
nest-asyncio==1.6.0
numpy==2.2.0
orjson==3.10.12
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
protobuf==5.29.2
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pycld2==0.41
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
streamlit==1.41.1
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
